// src/assets/types/constants.manufacturer.tsx
export enum constantsM {
    SIGNUP_STARTM = "SIGNUP_STARTM",
    SIGNUP_SUCCESSM = "SIGNUP_SUCCESSM",
    SIGNUP_ERRORM = "SIGNUP_FAILM",
    SIGNIN_STARTM = "SIGNIN_STARTM",
    SIGNIN_SUCCESSM = "SIGNIN_SUCCESSM",
    SIGNIN_ERRORM = "SIGNIN_FAILM",
    LOGOUTM = "LOGOUT",
    REFRESH_TOKENM = "REFRESH_TOKENM",

    CHANGE_PASSWORD_STARTM = "DELETE_STARTM",
    CHANGE_PASSWORD_SUCCESSM = "DELETE_SUCCESSM",
    CHANGE_PASSWORD_ERRORM = "DELETE_ERRORM",
    
    CREATE_STARTM = "CREATE_STARTM",
    CREATE_SUCCESSM = "CREATE_SUCCESSM",
    CREATE_ERRORM = "CREATE_ERRORM",
    CREATE_ALL_STARTM = "CREATE_STARTM",
    CREATE_ALL_SUCCESSM = "CREATE_SUCCESSM",
    CREATE_ALL_ERRORM = "CREATE_ERRORM",
    RETRIEVE_STARTM = "RETRIEVE_STARTM",
    RETRIEVE_SUCCESSM = "RETRIEVE_SUCCESSM",
    RETRIEVE_ERRORM = "RETRIEVE_ERRORM",
    RETRIEVE_ALL_STARTM = "RETRIEVE_ALL_STARTM",
    RETRIEVE_ALL_SUCCESSM = "RETRIEVE_ALL_SUCCESSM",
    RETRIEVE_ALL_ERRORM = "RETRIEVE_ALL_ERRORM",
    UPDATE_STARTM = "UPDATE_STARTM",
    UPDATE_SUCCESSM = "UPDATE_SUCCESSM",
    UPDATE_ERRORM = "UPDATE_ERRORM",
    DELETE_STARTM = "DELETE_STARTM",
    DELETE_SUCCESSM = "DELETE_SUCCESSM",
    DELETE_ERRORM = "DELETE_ERRORM",
    DELETE_ALL_STARTM = "DELETE_ALL_STARTM",
    DELETE_ALL_SUCCESSM = "DELETE_ALL_SUCCESSM",
    DELETE_ALL_ERRORM = "DELETE_ALL_ERRORM",
}